requirements.txt
exchangelib==4.6.1
requests==2.26.0
pyyaml==6.0